requirements.txt
-i https://pypi.org/simple
amqp==5.1.1; python_version >= '3.6'
asgiref==3.7.2; python_version >= '3.7'
async-timeout==4.0.3; python_full_version <= '3.11.2'
billiard==4.1.0; python_version >= '3.7'
celery==5.3.4; python_version >= '3.8'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
cron-descriptor==1.4.0
django==4.2.7; python_version >= '3.8'
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-taggit==5.0.1; python_version >= '3.8'
django-timezone-field==6.0.1; python_version >= '3.8' and python_version < '4.0'
djangorestframework==3.14.0; python_version >= '3.6'
idna==3.4; python_version >= '3.5'
kombu==5.3.2; python_version >= '3.8'
prompt-toolkit==3.0.39; python_full_version >= '3.7.0'
python-crontab==3.0.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
redis==5.0.1; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.8.0; python_version < '3.11'
tzdata==2023.3; python_version >= '2'
urllib3==2.0.7; python_version >= '3.7'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.9
